The Fireball Spell

+++ part_9/src/ecs/id.rs
@@ -1,6 +1,6 @@
 /// A unique type for `Entity` IDs. This is just here to make sure it's easy to change
 /// to another actual primitive type if necessary.
-pub type EntityId = u8;
+pub type EntityId = u16;
 
 /// Generator for IDs which are used to identify an `Entity`
 pub struct IdGenerator {
